requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.0.4
django-livereload-server==0.4
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
django-livereload-server==0.4
djangorestframework==3.13.1
idna==3.3
pycodestyle==2.8.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
urllib3==1.26.9
